requirements.txt
streamlit
ortools
google-genai
plotly
pandas
numpy